Restructured PID code to match the structure before
