+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Helper/Intake.java
@@ -1,5 +1,6 @@
 package org.firstinspires.ftc.teamcode.Helper;
 
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.Servo;
@@ -9,15 +10,16 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 public class Intake {
 
     public CRServo servo;
-    private ElapsedTime runtime = new ElapsedTime();
+    public LinearOpMode myOpMode;
 
-    //Init hardware map
-    HardwareMap hwMap = null;
+    //Constructor
+    public Intake(LinearOpMode opmode) {
+        myOpMode = opmode;
+    }
 
     //Initializing hardware maps and motors for intake.
-    public void init(HardwareMap ahwMap) {
-        hwMap = ahwMap;
-        servo = hwMap.get(CRServo.class, "intake");
+    public void init() {
+        servo = myOpMode.hardwareMap.get(CRServo.class, "intake");
     }
 
     public void MoveIntake(double position) {
