+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Helper/ProportionalControl.java
@@ -1,0 +1,112 @@
+package org.firstinspires.ftc.teamcode.Helper;
+
+//****************************************************************************************************
+
+import com.qualcomm.robotcore.util.ElapsedTime;
+import com.qualcomm.robotcore.util.Range;
+
+/***
+ * This class is used to implement a proportional controller which can calculate the desired output power
+ * to get an axis to the desired setpoint value.
+ * It also implements an acceleration limit, and a max power output.
+ */
+class ProportionalControl {
+    double  lastOutput;
+    double  gain;
+    double  accelLimit;
+    double  defaultOutputLimit;
+    double  liveOutputLimit;
+    double  setPoint;
+    double  tolerance;
+    double deadband;
+    boolean circular;
+    boolean inPosition;
+    ElapsedTime cycleTime = new ElapsedTime();
+
+    public ProportionalControl(double gain, double accelLimit, double outputLimit, double tolerance, double deadband, boolean circular) {
+        this.gain = gain;
+        this.accelLimit = accelLimit;
+        this.defaultOutputLimit = outputLimit;
+        this.liveOutputLimit = outputLimit;
+        this.tolerance = tolerance;
+        this.deadband = deadband;
+        this.circular = circular;
+        reset(0.0);
+    }
+
+    /**
+     * Determines power required to obtain the desired setpoint value based on new input value.
+     * Uses proportional gain, and limits rate of change of output, as well as max output.
+     * @param input  Current live control input value (from sensors)
+     * @return desired output power.
+     */
+    public double getOutput(double input) {
+        double error = setPoint - input;
+        double dV = cycleTime.seconds() * accelLimit;
+        double output;
+
+        // normalize to +/- 180 if we are controlling heading
+        if (circular) {
+            while (error > 180)  error -= 360;
+            while (error <= -180) error += 360;
+        }
+
+        inPosition = (Math.abs(error) < tolerance);
+
+        // Prevent any very slow motor output accumulation
+        if (Math.abs(error) <= deadband) {
+            output = 0;
+        } else {
+            // calculate output power using gain and clip it to the limits
+            output = (error * gain);
+            output = Range.clip(output, -liveOutputLimit, liveOutputLimit);
+
+            // Now limit rate of change of output (acceleration)
+            if ((output - lastOutput) > dV) {
+                output = lastOutput + dV;
+            } else if ((output - lastOutput) < -dV) {
+                output = lastOutput - dV;
+            }
+        }
+
+        lastOutput = output;
+
+        return output;
+    }
+
+    public boolean inPosition(){
+        return inPosition;
+    }
+    public double getSetpoint() {return setPoint;}
+
+    /**
+     * Saves a new setpoint and resets the output power history.
+     * This call allows a temporary power limit to be set to override the default.
+     * @param setPoint
+     * @param powerLimit
+     */
+    public void reset(double setPoint, double powerLimit) {
+        liveOutputLimit = Math.abs(powerLimit);
+        this.setPoint = setPoint;
+        reset();
+    }
+
+    /**
+     * Saves a new setpoint and resets the output power history.
+     * @param setPoint
+     */
+    public void reset(double setPoint) {
+        liveOutputLimit = defaultOutputLimit;
+        this.setPoint = setPoint;
+        reset();
+    }
+
+    /**
+     * Leave everything else the same, Just restart the acceleration timer and set output to 0
+     */
+    public void reset() {
+        cycleTime.reset();
+        inPosition = false;
+        lastOutput = 0.0;
+    }
+}
