+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Helper/Robot.java
@@ -1,5 +1,6 @@
 package org.firstinspires.ftc.teamcode.Helper;
 
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
@@ -11,52 +12,26 @@ public class Robot {
     Properties that describe hardware.
      */
     private ElapsedTime runtime = new ElapsedTime();
-    double timeout_ms = 0;
 
-    //Vision
-
-//
-//    public static final boolean USE_WEBCAM = true;  // true for webcam, false for phone camera
-//
-//    /**
-//     * {@link #aprilTag} is the variable to store our instance of the AprilTag processor.
-//     */
-//    public AprilTagProcessor aprilTag;
-//
-//    /**
-//     * {@link #tfod} is the variable to store our instance of the TensorFlow Object Detection processor.
-//     */
-//    public TfodProcessor tfod;
-//
-//    /**
-//     * {@link #myVisionPortal} is the variable to store our instance of the vision portal.
-//     */
-//    public VisionPortal myVisionPortal;
-//
+    private LinearOpMode myOpMode;
 
 
-    public Chassis chassis = new Chassis();
-    public Intake intake = new Intake();
-    public Arm arm = new Arm();
-    public Wrist wrist = new Wrist();
-    public Claw claw = new Claw();
+    //Contructor for robot class
+    public Robot(LinearOpMode opMode) {
+        myOpMode = opMode;
+    }
 
+    public Chassis chassis = new Chassis(myOpMode);
+    public Intake intake = new Intake(myOpMode);
+    public Arm arm = new Arm(myOpMode);
+    public Wrist wrist = new Wrist(myOpMode);
+    public Claw claw = new Claw(myOpMode);
 
-    /* local OpMode members. */
-    //Init hardware map
-    HardwareMap hwMap = null;
-
-
-    public ElapsedTime period = new ElapsedTime();
-    //tells you how long the robot has run for
-
-
-    public void init(HardwareMap ahwMap) throws InterruptedException {
-        hwMap = ahwMap;
-        chassis.init(hwMap);
-        intake.init(hwMap);
-        arm.init(hwMap);
-        wrist.init(hwMap);
-        claw.init(hwMap);
+    public void init() throws InterruptedException {
+        chassis.init(true);
+        intake.init();
+        arm.init();
+        wrist.init();
+        claw.init();
     }
 }
